+++ pyrealm/pmodel/pmodel.py
@@ -12,8 +12,12 @@ from numpy.typing import NDArray
 
 from pyrealm.constants import CoreConst, PModelConst
 from pyrealm.core.utilities import check_input_shapes, summarize_attrs
-from pyrealm.pmodel.functions import calc_ftemp_inst_rd, calc_modified_arrhenius_factor
-from pyrealm.pmodel.jmax_limitation import JmaxLimitation
+from pyrealm.pmodel.arrhenius import ARRHENIUS_METHOD_REGISTRY, ArrheniusFactorABC
+from pyrealm.pmodel.functions import calc_ftemp_inst_rd
+from pyrealm.pmodel.jmax_limitation import (
+    JMAX_LIMITATION_CLASS_REGISTRY,
+    JmaxLimitationABC,
+)
 from pyrealm.pmodel.optimal_chi import OPTIMAL_CHI_CLASS_REGISTRY, OptimalChiABC
 from pyrealm.pmodel.pmodel_environment import PModelEnvironment
 from pyrealm.pmodel.quantum_yield import QUANTUM_YIELD_CLASS_REGISTRY, QuantumYieldABC
@@ -31,7 +35,8 @@ class PModel:
        :math:`\ce{CO2}` partial pressure ratios (:math:`\chi`), using one of the
        methods based on :class:`~pyrealm.pmodel.optimal_chi.OptimalChiABC`.
     2. Estimate limitation factors to :math:`V_{cmax}` and :math:`J_{max}` using
-       :class:`~pyrealm.pmodel.jmax_limitation.JmaxLimitation`.
+       one of the methods implemented using 
+       :class:`~pyrealm.pmodel.jmax_limitation.JmaxLimitationABC`.
     3. Optionally, estimate productivity measures including GPP by supplying FAPAR and
        PPFD using the :meth:`~pyrealm.pmodel.pmodel.PModel.estimate_productivity`
        method.
@@ -50,7 +55,7 @@ class PModel:
             \text{LUE} = \phi_0 \cdot m_j \cdot f_v \cdot M_C
 
       where :math:`f_v` is a limitation factor defined in
-      :class:`~pyrealm.pmodel.jmax_limitation.JmaxLimitation` and :math:`M_C` is the
+      :class:`~pyrealm.pmodel.jmax_limitation.JmaxLimitationABC` and :math:`M_C` is the
       molar mass
       of carbon.
 
@@ -77,13 +82,12 @@ class PModel:
             \]
 
     where  :math:`f_v, f_j` are limitation terms described in
-    :class:`~pyrealm.pmodel.jmax_limitation.JmaxLimitation`
+    :class:`~pyrealm.pmodel.jmax_limitation.JmaxLimitationABC`
 
-    * The maximum carboxylation capacity (mol C m-2) normalised to the standard
-      temperature as: :math:`V_{cmax25} = V_{cmax}  / fv(t)`, where :math:`fv(t)` is the
-      instantaneous temperature response of :math:`V_{cmax}` calculated using a modified
-      Arrhenius equation
-      :func:`~pyrealm.pmodel.functions.calc_modified_arrhenius_factor`.
+    * The values :math:`V_{cmax25}` and :math:`J_{max25}` are the expected values of 
+      :math:`V_{cmax}` and :math:`J_{cmax}` normalised to the standard temperature
+      (25°C). The rates are normalised using an Arrhenius scaling factor using the
+      specified `method_arrhenius`.
 
     * Dark respiration, calculated as:
 
@@ -168,6 +172,7 @@ class PModel:
         method_kphio: str = "temperature",
         method_optchi: str = "prentice14",
         method_jmaxlim: str = "wang17",
+        method_arrhenius: str = "simple",
         reference_kphio: float | NDArray | None = None,
     ):
         self.shape: tuple = env.shape
@@ -220,15 +225,31 @@ class PModel:
         """A subclass of QuantumYieldABC, providing kphio calculation."""
 
         # -----------------------------------------------------------------------
+        # Set up the calculation of Arrhenius scaling
+        # -----------------------------------------------------------------------
+        if method_arrhenius not in ARRHENIUS_METHOD_REGISTRY:
+            raise ValueError(f"Unknown Arrhenius scaling method: {method_arrhenius}")
+
+        self.method_arrhenius: str = method_arrhenius
+        """The method used to calculate Arrhenius factors.
+        
+        We currently strongly recommend the use of the default `simple` method for this
+        setting. 
+        """
+
+        # -----------------------------------------------------------------------
         # Calculation of Jmax limitation terms
         # -----------------------------------------------------------------------
+        if method_jmaxlim not in JMAX_LIMITATION_CLASS_REGISTRY:
+            raise ValueError(f"Unknown Jmax limitation method: {method_jmaxlim}")
+
         self.method_jmaxlim: str = method_jmaxlim
         """Records the method used to calculate Jmax limitation."""
 
-        self.jmaxlim: JmaxLimitation = JmaxLimitation(
-            self.optchi, method=self.method_jmaxlim, pmodel_const=self.pmodel_const
-        )
-        """Details of the Jmax limitation calculation for the model"""
+        self.jmaxlim: JmaxLimitationABC = JMAX_LIMITATION_CLASS_REGISTRY[
+            method_jmaxlim
+        ](optchi=self.optchi, pmodel_const=self.pmodel_const)
+        """The Jmax limitation terms for the model"""
 
         # -----------------------------------------------------------------------
         # Store the two efficiency predictions
@@ -260,8 +281,9 @@ class PModel:
         # -----------------------------------------------------------------------
         self._vcmax: NDArray[np.float64]
         self._vcmax25: NDArray[np.float64]
-        self._rd: NDArray[np.float64]
         self._jmax: NDArray[np.float64]
+        self._jmax25: NDArray[np.float64]
+        self._rd: NDArray[np.float64]
         self._gpp: NDArray[np.float64]
         self._gs: NDArray[np.float64]
         self._ppfd: NDArray[np.float64]
@@ -306,6 +328,12 @@ class PModel:
         """Maximum rate of electron transport (µmol m-2 s-1)."""
         self._check_estimated("jmax")
         return self._jmax
+
+    @property
+    def jmax25(self) -> NDArray[np.float64]:
+        """Maximum rate of electron transport at standard temperature (µmol m-2 s-1)."""
+        self._check_estimated("jmax25")
+        return self._jmax25
 
     @property
     def gs(self) -> NDArray[np.float64]:
@@ -370,25 +398,28 @@ class PModel:
         # GPP
         self._gpp = self.lue * iabs
 
-        # V_cmax
+        # Calculate V_cmax and J_max
         self._vcmax = self.kphio.kphio * iabs * self.optchi.mjoc * self.jmaxlim.f_v
+        self._jmax = 4 * self.kphio.kphio * iabs * self.jmaxlim.f_j
 
-        # Calculate the modified arrhenius factor to normalise V_cmax to V_cmax25
-        # - Get parameters
-        kk_a, kk_b, kk_ha, kk_hd = self.pmodel_const.kattge_knorr_kinetics
-        # Calculate entropy as a function of temperature _in °C_
-        kk_deltaS = kk_a + kk_b * self.env.tc
-        # Calculate the arrhenius factor
-        ftemp25_inst_vcmax = calc_modified_arrhenius_factor(
-            tk=self.env.tc + self.core_const.k_CtoK,
-            Ha=kk_ha,
-            Hd=kk_hd,
-            tk_ref=self.pmodel_const.plant_T_ref + self.core_const.k_CtoK,
-            mode=self.pmodel_const.modified_arrhenius_mode,
-            deltaS=kk_deltaS,
+        # Calculate V_cmax25 and J_max25
+        # - get an instance of the requested Arrhenius scaling method
+        arrh_factor_instance: ArrheniusFactorABC = ARRHENIUS_METHOD_REGISTRY[
+            self.method_arrhenius
+        ](
+            env=self.env,
+            reference_temperature=self.pmodel_const.plant_T_ref,
             core_const=self.core_const,
         )
+
+        # - Calculate and apply the scaling factors.
+        ftemp25_inst_vcmax = arrh_factor_instance.calculate_arrhenius_factor(
+            coefficients=self.pmodel_const.arrhenius_vcmax
+        )
         self._vcmax25 = self._vcmax / ftemp25_inst_vcmax
+        self._jmax25 = self._jmax / arrh_factor_instance.calculate_arrhenius_factor(
+            coefficients=self.pmodel_const.arrhenius_jmax
+        )
 
         # Dark respiration at growth temperature
         ftemp_inst_rd = calc_ftemp_inst_rd(self.env.tc, pmodel_const=self.pmodel_const)
@@ -397,9 +428,6 @@ class PModel:
             * (ftemp_inst_rd / ftemp25_inst_vcmax)
             * self._vcmax
         )
-
-        # Calculate Jmax
-        self._jmax = 4 * self.kphio.kphio * iabs * self.jmaxlim.f_j
 
         # AJ and AC
         a_j = self.kphio.kphio * iabs * self.optchi.mj * self.jmaxlim.f_v