+++ tests/unit/pmodel/test_functions_new.py
@@ -8,13 +8,6 @@ import numpy as np
 from numpy.testing import assert_allclose
 
 
-def test_calc_ftemp_arrh(tk=np.array([300]), ha=40000):
-    """Test calc_ftemp_arrh."""
-    from pyrealm.pmodel.functions import calc_ftemp_arrh
-
-    assert_allclose(calc_ftemp_arrh(tk, ha), 1.104622629609139)
-
-
 def test_calc_ftemp_inst_rd(tc=np.array([30.0])):
     """Test calc_ftemp_inst_rd."""
     from pyrealm.pmodel.functions import calc_ftemp_inst_rd